# api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentDoel": {
        "properties": {
          "artikel": {
            "title": "Artikel",
            "type": "string"
          },
          "bwbId": {
            "title": "Bwbid",
            "type": "string"
          },
          "lid": {
            "default": "",
            "title": "Lid",
            "type": "string"
          }
        },
        "required": [
          "bwbId",
          "artikel"
        ],
        "title": "AgentDoel",
        "type": "object"
      },
...
      "Bericht": {
        "description": "E\u00e9n beurt in het gesprek. `annotatie_titel` is het leesbare label van het\nannotatiedocument op het moment van de beurt \u2014 het bericht beschrijft zichzelf, zodat het\ngesprek leesbaar blijft als het document later verwijderd wordt (geen foreign key).",
        "properties": {
          "aangemaakt": {
            "default": "",
            "title": "Aangemaakt",
            "type": "string"
          },
          "annotatie_slug": {
            "default": "",
            "title": "Annotatie Slug",
            "type": "string"
          },
          "annotatie_titel": {
            "default": "",
            "title": "Annotatie Titel",
            "type": "string"
          },
          "bronnen": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Bronnen",
            "type": "array"
          },
          "denk": {
            "default": "",
            "title": "Denk",
            "type": "string"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "ontbrekend": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Ontbrekend",
            "type": "array"
          },
          "rol": {
            "$ref": "#/components/schemas/Rol"
          },
          "run_id": {
            "default": "",
            "title": "Run Id",
            "type": "string"
          },
          "tekst": {
            "default": "",
            "title": "Tekst",
            "type": "string"
          }
        },
        "required": [
          "rol"
        ],
        "title": "Bericht",
        "type": "object"
      },
...
      "ChatRequest": {
        "properties": {
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "doel": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentDoel"
              },
              {
                "type": "null"
              }
            ]
          },
          "question": {
            "default": "",
            "title": "Question",
            "type": "string"
          },
          "werkgebied": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Werkgebied"
          }
        },
        "title": "ChatRequest",
        "type": "object"
      },
...
      "Gesprek": {
        "properties": {
          "aangemaakt": {
            "default": "",
            "title": "Aangemaakt",
            "type": "string"
          },
          "berichten": {
            "items": {
              "$ref": "#/components/schemas/Bericht"
            },
            "title": "Berichten",
            "type": "array"
          },
          "bijgewerkt": {
            "default": "",
            "title": "Bijgewerkt",
            "type": "string"
          },
          "gebruiker": {
            "title": "Gebruiker",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "titel": {
            "default": "",
            "title": "Titel",
            "type": "string"
          }
        },
        "required": [
          "id",
          "gebruiker"
        ],
        "title": "Gesprek",
        "type": "object"
      },
      "GesprekAanmaken": {
        "properties": {
          "titel": {
            "default": "",
            "title": "Titel",
            "type": "string"
          }
        },
        "title": "GesprekAanmaken",
        "type": "object"
      },
...
      "GesprekSamenvatting": {
        "description": "Lichte lijst-weergave voor een gesprekgeschiedenis-overzicht.",
        "properties": {
          "aantal_berichten": {
            "default": 0,
            "title": "Aantal Berichten",
            "type": "integer"
          },
          "bijgewerkt": {
            "default": "",
            "title": "Bijgewerkt",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "titel": {
            "default": "",
            "title": "Titel",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "GesprekSamenvatting",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Rol": {
        "enum": [
          "user",
          "assistant"
        ],
        "title": "Rol",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "wetsanalyse-api (referentie-implementatie)",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/chat": {
      "post": {
        "operationId": "chat_v1_chat_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat",
        "tags": [
          "chat"
        ]
      }
    },
...
    "/v1/gesprekken": {
      "get": {
        "operationId": "lijst_gesprekken_v1_gesprekken_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GesprekSamenvatting"
                  },
                  "title": "Response Lijst Gesprekken V1 Gesprekken Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lijst Gesprekken",
        "tags": [
          "gesprekken"
        ]
      },
      "post": {
        "operationId": "maak_gesprek_v1_gesprekken_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GesprekAanmaken"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Gesprek"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Maak Gesprek",
        "tags": [
          "gesprekken"
        ]
      }
    },
...
  }
}